improvement(server): merge message-transformers into one session history module
